Add Critique model and update seeding and relations
This commit introduces the new Critique model as well as creating the corresponding 'critiques' table through a migration. The 'GenreSeeder.php' was also slightly adjusted, reducing the number of genre creations from 10 to 5. Meanwhile, the relation method in 'Book.php' was specified to BelongsToMany. Lastly, a .gitignore file was added to the 'debugbar' directory in 'storage'.