Add user authentication and login page
The commit includes the creation of a new UserController for handling user logins alongside a login blade view. A new route for the login page was added as well. An auth check was applied in the BookController to redirect non-authenticated users to the login page. Removed the welcome view which is no longer required.