Add Debugbar support and update models
This commit includes the integration of Debugbar for better development debugging. It also adds additional methods and properties to the Critique and Collection models. A new view file, "books.blade.php" and a new metadata file for PHPStorm, ".phpstorm.meta.php", have also been created. The update will facilitate the retrieval and display of book-related data.