Add functionality to update reading progression
A new functionality has been added to update the reading progression of a book. A new route, /updateProgression, has been added to handle POST requests, along with an updateProgression method in the BookController. The view 'book.book' was updated to include a form for setting the progression and an AJAX request to submit it.